This commit was manufactured by cvs2git to create tag 'glite-lb_R_1_9_3_1'.
Sprout from branch_RC31_3 2009-10-15 14:24:23 UTC Zdeněk Šustr <sustr4@cesnet.cz> 'Module repacked'
Cherrypick from branch_RC31_3 2008-06-10 10:50:05 UTC Zdeněk Šustr <sustr4@cesnet.cz> 'push minor version':
    org.glite.lb.utils/project/version.properties
Cherrypick from branch_RC31_3 2009-11-18 14:12:37 UTC Zdeněk Šustr <sustr4@cesnet.cz> 'New LB version':
    org.glite.lb/project/version.properties
Cherrypick from master 2009-11-14 14:50:42 UTC František Dvořák <valtri@civ.zcu.cz> 'Finalize build.':
    org.glite.lb.harvester/Makefile
    org.glite.lb.harvester/configure
    org.glite.lb.harvester/doc/INSTALL
    org.glite.lb.harvester/doc/README
    org.glite.lb.harvester/examples/test.sh
    org.glite.lb.harvester/examples/test.sql
    org.glite.lb.harvester/project/ChangeLog
    org.glite.lb.harvester/project/package.description
    org.glite.lb.harvester/project/package.summary
    org.glite.lb.harvester/project/version.properties
    org.glite.lb.harvester/src/harvester.c
    org.glite.lb.logger/config/startup
    org.glite.lb.logger/configure
    org.glite.lb.logger/project/ChangeLog
    org.glite.lb.logger/project/package.description
    org.glite.lb.logger/project/package.summary
    org.glite.lb.logger/src-nt/Connection.H
    org.glite.lb.logger/src-nt/Connection.cpp
    org.glite.lb.logger/src-nt/EventManager.H
    org.glite.lb.logger/src-nt/EventManager.cpp
    org.glite.lb.logger/src-nt/Exception.H
    org.glite.lb.logger/src-nt/HTTPTransport.H
    org.glite.lb.logger/src-nt/HTTPTransport.cpp
    org.glite.lb.logger/src-nt/InputChannel.H
    org.glite.lb.logger/src-nt/InputChannel.cpp
    org.glite.lb.logger/src-nt/Makefile
    org.glite.lb.logger/src-nt/Message.H
    org.glite.lb.logger/src-nt/MessageStore.H
    org.glite.lb.logger/src-nt/MessageStore.cpp
    org.glite.lb.logger/src-nt/PlainConnection.H
    org.glite.lb.logger/src-nt/PlainConnection.cpp
    org.glite.lb.logger/src-nt/PluginManager.H
    org.glite.lb.logger/src-nt/PluginManager.cpp
    org.glite.lb.logger/src-nt/Properties.H
    org.glite.lb.logger/src-nt/SocketInput.cpp
    org.glite.lb.logger/src-nt/ThreadPool.H
    org.glite.lb.logger/src-nt/ThreadPool.cpp
    org.glite.lb.logger/src-nt/Transport.H
    org.glite.lb.logger/src-nt/Transport.cpp
    org.glite.lb.logger/src-nt/main.cpp
    org.glite.lb.logger/src-nt/test/EventManagerTest.cpp
    org.glite.lb.logger/src-nt/test/ThreadPoolTest.cpp
    org.glite.lb.logger/src/event_store_http.c
    org.glite.lb.logger/src/http.c
    org.glite.lb.logger/src/input_queue_socket_http.c
    org.glite.lb.logger/src/queue_mgr_http.c
    org.glite.lb.logger/src/send_event_http.c
    org.glite.lb.logger/src/server_msg_http.c
Cherrypick from master 2009-03-25 15:09:32 UTC Michal Voců <michal@ruk.cuni.cz> 'set socket option':
    org.glite.lb.logger/doc/glite-lb-interlogd.8
    org.glite.lb.logger/doc/glite-lb-logd.8
    org.glite.lb.logger/src/il_error.c
    org.glite.lb.logger/src/il_error.h
    org.glite.lb.logger/src/input_queue_socket.c
    org.glite.lb.logger/src/logd.c
    org.glite.lb.logger/src/logd_proto.c
    org.glite.lb.logger/src/perftest_il.sh
    org.glite.lb.logger/src/queue_thread.c
Cherrypick from branch_2_0_notif_GL31 2009-04-14 11:19:19 UTC Aleš Křenek <ljocha@ics.muni.cz> 'Handle copying messages with empty destination':
    org.glite.lb.logger/Makefile
    org.glite.lb.logger/src/event_queue.c
    org.glite.lb.logger/src/event_store.c
    org.glite.lb.logger/src/il_master.c
    org.glite.lb.logger/src/interlogd.c
    org.glite.lb.logger/src/interlogd.h
    org.glite.lb.logger/src/queue_mgr.c
    org.glite.lb.logger/src/recover.c
    org.glite.lb.logger/src/send_event.c
    org.glite.lb.logger/src/server_msg.c

+++ org.glite.lb.logger/src-nt/SocketInput.cpp
@@ -6,7 +6,8 @@
 
 #include "ThreadPool.H"
 #include "SocketInput.H"
-
+#include "InputChannel.H"
+#include "Exception.H"
 
 
 // create unix domain socket for input
@@ -21,6 +22,7 @@ SocketInput::SocketInput(const char *path,
 	saddr.sun_family = AF_UNIX;
 	strcpy(saddr.sun_path, path);
 	fd = socket(PF_UNIX, SOCK_STREAM, 0);
+	if(fd < 0) throw new Exception;
 	if(connect(fd, (struct sockaddr*)&saddr, sizeof(saddr.sun_path)) < 0) {
 		if(errno == ECONNREFUSED) {
 			unlink(saddr.sun_path);
@@ -29,9 +31,10 @@ SocketInput::SocketInput(const char *path,
 		// another instance running
 		// throw new Exception
 	}
-	bind(fd, (struct sockaddr *)&saddr, sizeof(saddr));
-	listen(fd, SOCK_QUEUE_MAX);
-	ThreadPool::instance()->setWorkAccept(this);
+	if(bind(fd, (struct sockaddr *)&saddr, sizeof(saddr)) < 0) 
+		throw new Exception;
+	if(listen(fd, SOCK_QUEUE_MAX) < 0)
+		throw new Exception;
 }
 
 
@@ -48,8 +51,9 @@ void
 SocketInput::onReady()
 {
 	Connection *conn = cFactory->accept(fd);
-	Transport  *trans = tFactory->newTransport(conn);
-	ThreadPool::instance()->queueWorkRead(trans);
+	Transport  *trans = tFactory->newTransport();
+	InputChannel *channel = new InputChannel(conn, trans);
+	channel->start();
 }
